+++ erik/2-mysql/3-full_friends/server.py
@@ -16,7 +16,7 @@ def index():
     friends = mysql.query_db("SELECT * FROM friends")
     print "many friends ",friends
     # print friends
-    return render_template('index.html',all_friends=friends)
+    return render_template('index.html',friends=friends)
 
 # /friends
 # Create a new friend
@@ -33,39 +33,68 @@ def create():
 
 # /friends/<friend_id>
 # Show a single friend by id
-@app.route('/friends/<friend_id>')
-def show(friend_id):
-    # to insert param variable :variable_name
-    query = "SELECT * FROM friends WHERE id = :specific_id"
-    # Then define a dictionary with key that matches :variable_name in query.
-    print "friend data: ", friend_id
-    data = {'specific_id': friend_id}
-    print "query data:",data
-    # Execute query, passing in a parameter obj
-    friends = mysql.query_db(query, data)
-    # pass expected parameter(s) to the page to be rendered
-    return render_template('index.html', all_friends=friends)
+@app.route('/friends/<id>',methods=['POST','GET'])
+def show(id):
+    # POST it is an update
+    if request.method == "POST":
+        query = "UPDATE friends SET first_name = :first_name, last_name = :last_name, occupation = :occupation WHERE id = :id"
+        data = {
+                 'first_name':  request.form['first_name'],
+                 'last_name':   request.form['last_name'],
+                 'occupation':  request.form['occupation'],
+                 'id':          request.form['id']
+               }
+        mysql.query_db(query, data)
+        return redirect('/')
 
-# /update_friend/<friend_id>
+    # GET it is a show a single friend
+    else:
+        # to insert param variable :variable_name
+        query = "SELECT * FROM friends WHERE id = :id"
+        # Then define a dictionary with key that matches :variable_name in query.
+        print "friend data: ", id
+        data = {'id': id}
+        print "query data:",data
+        # Execute query, passing in a parameter obj
+        friends = mysql.query_db(query, data)
+        # pass expected parameter(s) to the page to be rendered
+        return render_template('index.html', friends=friends)
+
+
 # Update a friend by id
-@app.route('/update_friend/<friend_id>', methods=['POST'])
-def update(friend_id):
-    query = "UPDATE friends SET first_name = :first_name, last_name = :last_name, occupation = :occupation WHERE id = :id"
-    data = {
-             'first_name': request.form['first_name'],
-             'last_name':  request.form['last_name'],
-             'occupation': request.form['occupation'],
-             'id': friend_id
-           }
-    mysql.query_db(query, data)
-    return redirect('/')
+# /friends/<id>/edit
+@app.route('/friends/<id>/edit', methods=['POST','GET'])
+def update(id):
 
-# /unfriend/<friend_id>
-# Remove a friend by id
-@app.route('/remove_friend/<friend_id>', methods=['POST'])
-def delete(friend_id):
+    # POST it is an update
+    if request.method == "POST":
+        # grab the post info and display it on the page
+        friend={
+            'first_name':  request.form['first_name'],
+            'last_name':   request.form['last_name'],
+            'occupation':  request.form['occupation'],
+            'id':          request.form['id']
+        }
+        return render_template('edit.html', friend=friend)
+    # GET the information and ready the page for editing
+    else:
+        # Get the info, render the edit page
+        query = "SELECT * FROM friends WHERE id = :id"
+        # Then define a dictionary with key that matches :variable_name in query.
+        print "friend data: ", id
+        data = {'id': id}
+        print "query data:",data
+        # Execute query, passing in a parameter obj
+        friends = mysql.query_db(query, data)
+        # pass expected parameter(s) to the page to be rendered
+        return render_template('edit.html', friend=friend)
+
+# /friends/<id>/delete
+# Delete a friend by id
+@app.route('/friends/<id>/delete', methods=['POST'])
+def delete(id):
     query = "DELETE FROM friends WHERE id = :id"
-    data = {'id': friend_id}
+    data = {'id': id}
     mysql.query_db(query, data)
     return redirect('/')
