+++ erik/2-mysql/3-full_friends/templates/index.html
@@ -3,22 +3,31 @@
   <title>Friends</title>
 </head>
 <body>
-  <h1>These are all my friends!</h1>
-  {% for friend in all_friends %}
-  <p>First ID: {{ friend['id'] }}
-  <p>First Name: {{ friend['first_name'] }}
-  <p>Last Name: {{ friend['last_name'] }}
 
-  <p>Occupation: {{ friend['occupation'] }}
-      <hr>
+   <h1>Friends</h1>
+
+  <h2>These are all my friends!</h2>
+  {% for friend in friends %}
+      <form action='/friends/{{ friend['id'] }}/edit' method='post'>
+         <p>ID: {{ friend['id'] }}</p>
+         <p>First Name: {{ friend['first_name'] }}</p>
+         <p>Last Name: {{ friend['last_name'] }}</p>
+         <p>Occupation: {{ friend['occupation'] }}</p>
+         <input type='hidden' name='id' value='{{ friend['id']}}' readonly>
+         <button type="submit" value="Edit">Edit Friend</button>
+      </form>
+      <form action='/friends/{{ friend['id'] }}/delete' method='post'>
+         <button type="submit" value="Delete">Delete Ex-Friend</button>
+         <hr>
+      </form>
    {% endfor %}
 
-  <h2>Add a Friend</h2>
-  <form action='/friends' method='POST'>
-    <label for="first_name">First Name:<input type="text" name="first_name" id="first_name"></label>
-    <label for="last_name">Last Name:<input type="text" name="last_name" id="last_name"></label>
-    <label for="occupation">Occupation:<input type="text" name="occupation" id="occupation"></label>
-    <input type="submit" value="Add">
-  </form>
+  <h2>Add A Friend</h2>
+   <form action='/friends' method='post'>
+      <label for="first_name">First Name:<input type="text" name="first_name" id="first_name"></label>
+      <label for="last_name">Last Name:<input type="text" name="last_name" id="last_name"></label>
+      <label for="occupation">Occupation:<input type="text" name="occupation" id="occupation"></label>
+      <input type="submit" value="Add">
+   </form>
 </body>
 </html>
